fixing bugs and adding todos

+++ src/player/story_progress.rs
@@ -2,6 +2,6 @@ use serde::{ Deserialize, Serialize };
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct StoryProgress {
-    pub areas_completed: u8,
-    pub current_area_progress: u8,
+    pub areas_completed: usize,
+    pub current_area_progress: usize,
 }
